+++ client/src/components/ProductLaunchLogo.js
